Upgrade skynet-js and fix broken tests
Upgrade `skynet-js`. Doing so broke the tests so I fixed that.

The only relevant thing that changed in v4.2.0-beta is that we updated the axios
dependency, and it would seem like it behaves differently now. Previously, we
were partially mocking axios, but that didn't have any effect on the tests. Now,
however, it seems like the mock does not let through any actual network
requests, even if it is a partial mock. To fix this, all of the requests would
have to be mocked.

But I think it would be more valuable to let these tests keep making actual
network requests. I will remove the partial mock as it was actually a mistake
and not doing anything.