Fix `Concurrent access prevented in SkyDB` error after test shutdown
It looks like when a test is unexpectedly quit, any open locks for SkyDB V2
entries stay open.

I figured we could avoid this by using random keys per test, instead of a single
pair of keys for all tests.

It's not clear yet whether this is a problem in production as the community has
not reported it. But this PR will make CI test output easier to read and debug.